Export config as `govuk-frontend-config` npm workspace
