+++ app/package.json
@@ -24,6 +24,7 @@
     "govuk_frontend_toolkit": "^9.0.1",
     "govuk_template_jinja": "^0.26.0",
     "govuk-elements-sass": "3.1.3",
+    "govuk-frontend-config": "*",
     "gulp": "^4.0.2",
     "gulp-cli": "^2.3.0",
     "highlight.js": "^11.7.0",
