+++ app/src/start.mjs
@@ -1,4 +1,4 @@
-import { ports } from '../../config/index.js'
+import { ports } from 'govuk-frontend-config'
 
 import app from './app.mjs'
