+++ config/package.json
@@ -1,0 +1,20 @@
+{
+  "private": true,
+  "name": "govuk-frontend-config",
+  "description": "GOV.UK Frontend shared config",
+  "main": "index.js",
+  "engines": {
+    "node": "^18.12.0",
+    "npm": "^8.1.0 || ^9.1.0"
+  },
+  "license": "MIT",
+  "dependencies": {
+    "govuk-frontend": "*",
+    "jest-axe": "^7.0.1",
+    "jest-dev-server": "^8.0.5",
+    "jest-environment-jsdom": "^29.5.0",
+    "jest-environment-node-single-context": "^29.0.0",
+    "jest-environment-puppeteer": "^8.0.6",
+    "puppeteer": "^19.8.2"
+  }
+}
