+++ jest-dev-server.config.js
@@ -1,4 +1,4 @@
-const { ports } = require('./config/index.js')
+const { ports } = require('govuk-frontend-config')
 
 /**
  * @type {import('jest-dev-server').Config}
