+++ lib/file-helper.js
@@ -3,10 +3,9 @@ const { join, parse, relative } = require('path')
 
 const fm = require('front-matter')
 const { glob } = require('glob')
+const { paths } = require('govuk-frontend-config')
 const yaml = require('js-yaml')
 const { minimatch } = require('minimatch')
-
-const { paths } = require('../config')
 
 /**
  * Directory listing for path
