+++ lib/helper-functions.unit.test.mjs
@@ -1,6 +1,6 @@
 import { join, relative } from 'path'
 
-import { paths } from '../config/index.js'
+import { paths } from 'govuk-frontend-config'
 
 import {
   componentNameToClassName,
