+++ lib/jest-helpers.js
@@ -1,9 +1,8 @@
 const { join } = require('path')
 
+const { paths } = require('govuk-frontend-config')
 const { configureAxe } = require('jest-axe')
 const { compileAsync, compileStringAsync, Logger } = require('sass-embedded')
-
-const { paths } = require('../config/index.js')
 
 const sassPaths = [
   join(paths.src, 'govuk'),
