+++ lib/nunjucks-helpers.js
@@ -1,9 +1,8 @@
 const { join } = require('path')
 
+const { paths } = require('govuk-frontend-config')
 const nunjucks = require('nunjucks')
 const { outdent } = require('outdent')
-
-const { paths } = require('../config/index.js')
 
 const { componentNameToMacroName } = require('./helper-functions')
