+++ lib/puppeteer-helpers.js
@@ -1,4 +1,4 @@
-const { ports } = require('../config')
+const { ports } = require('govuk-frontend-config')
 
 const { componentNameToClassName } = require('./helper-functions')
 const { render } = require('./nunjucks-helpers')
