+++ package.json
@@ -9,6 +9,7 @@
   "license": "MIT",
   "workspaces": [
     "app",
+    "config",
     "docs/examples/*",
     "package",
     "tasks"
@@ -80,15 +81,13 @@
     "eslint-plugin-jsdoc": "^40.1.1",
     "eslint-plugin-n": "^15.7.0",
     "eslint-plugin-promise": "^6.1.1",
+    "govuk-frontend-config": "*",
     "govuk-frontend-tasks": "*",
     "html-validate": "7.14.0",
     "husky": "^8.0.3",
     "jest": "^29.5.0",
     "jest-axe": "^7.0.1",
-    "jest-dev-server": "^8.0.3",
-    "jest-environment-jsdom": "^29.5.0",
     "jest-environment-node-single-context": "^29.0.0",
-    "jest-environment-puppeteer": "^8.0.6",
     "jest-puppeteer": "^8.0.6",
     "jest-serializer-html": "^7.1.0",
     "lint-staged": "^13.2.0",
