+++ tasks/npm.mjs
@@ -1,7 +1,6 @@
 import runScript from '@npmcli/run-script'
+import { paths } from 'govuk-frontend-config'
 import PluginError from 'plugin-error'
-
-import { paths } from '../config/index.js'
 
 import { isDev } from './helpers/task-arguments.mjs'
 import { task } from './index.mjs'
