+++ tasks/package.json
@@ -16,6 +16,7 @@
     "chalk": "^5.2.0",
     "cpy": "^9.0.1",
     "del": "^7.0.0",
+    "govuk-frontend-config": "*",
     "gulp": "^4.0.2",
     "gulp-cli": "^2.3.0",
     "js-yaml": "^4.1.0",
