+++ tasks/styles.mjs
@@ -2,13 +2,13 @@ import { readFile } from 'fs/promises'
 import { join, parse } from 'path'
 
 import chalk from 'chalk'
+import { paths } from 'govuk-frontend-config'
 import PluginError from 'plugin-error'
 import postcss from 'postcss'
 // eslint-disable-next-line import/default
 import postcssrc from 'postcss-load-config'
 import { compileAsync } from 'sass-embedded'
 
-import { paths } from '../config/index.js'
 import { getListing } from '../lib/file-helper.js'
 
 import { assets } from './index.mjs'
